requirements.txt
# Python 3.12.4

# Data processing (Core dependencies)
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Configuration
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0

# Logging
python-json-logger>=2.0

# Utilities
python-dateutil>=2.8.0

# AWS integration
boto3>=1.42.40

# Visualization (for notebooks and dashboard)
plotly>=6.5.2
seaborn>=0.12.0
matplotlib>=3.7.0

# OpenAI integration (Phase 2)
openai>=2.16.0

# Excel support
openpyxl>=3.1.5

# Jupyter notebooks
jupyter>=1.1.1
ipykernel>=6.25.0

# Progress bars
tqdm>=4.65.0

# Environment variables
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning
tensorflow

# Hyperparameter tuning
Optuna